Move hasOnePreselectedDocument, enabledDocuments into utils
