+++ src/components/utils/index.js
@@ -164,3 +164,14 @@ export const capitalise = (string) => {
 
   return string
 }
+
+export const hasOnePreselectedDocument = (steps) =>
+  enabledDocuments(steps).length === 1
+
+export const enabledDocuments = (steps) => {
+  const documentStep = steps.find((step) => step.type === 'document')
+  const docTypes =
+    documentStep && documentStep.options && documentStep.options.documentTypes
+  console.log('docTypes:', docTypes)
+  return docTypes ? Object.keys(docTypes).filter((type) => docTypes[type]) : []
+}
